# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Hotel Review Insights API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /analytics/overview:
    get:
      operationId: getAnalyticsOverview
      tags: [analytics]
      summary: Dashboard KPI cards — avg rating, weekly change, total reviews, positive %
      parameters:
        - name: hotelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Overview metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsOverview"
  /analytics/rating-trend:
    get:
      operationId: getRatingTrend
      tags: [analytics]
      summary: Monthly average ratings per hotel over time
      parameters:
        - name: hotelId
          in: query
          schema:
            type: integer
        - name: months
          in: query
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Rating trend data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RatingTrendPoint"
...
  /analytics/sentiment-distribution:
    get:
      operationId: getSentimentDistribution
      tags: [analytics]
      summary: Positive, neutral, negative review counts per hotel
      parameters:
        - name: hotelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Sentiment distribution
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SentimentCount"
  /analytics/complaint-categories:
    get:
      operationId: getComplaintCategories
      tags: [analytics]
      summary: Complaint topic category counts
      parameters:
        - name: hotelId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Complaint categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ComplaintCategory"
...
components:
  schemas:
...
    AnalyticsOverview:
      type: object
      required: [averageRating, weeklyRatingChange, totalReviews, positivePercent, neutralPercent, negativePercent]
      properties:
        averageRating:
          type: number
        weeklyRatingChange:
          type: number
        totalReviews:
          type: number
        positivePercent:
          type: number
        neutralPercent:
          type: number
        negativePercent:
          type: number
    RatingTrendPoint:
      type: object
      required: [month, hotelId, hotelName, averageRating, reviewCount]
      properties:
        month:
          type: string
        hotelId:
          type: number
        hotelName:
          type: string
        averageRating:
          type: number
        reviewCount:
          type: number
...
    SentimentCount:
      type: object
      required: [hotelId, hotelName, positive, neutral, negative]
      properties:
        hotelId:
          type: number
        hotelName:
          type: string
        positive:
          type: number
        neutral:
          type: number
        negative:
          type: number
    ComplaintCategory:
      type: object
      required: [category, count, percent]
      properties:
        category:
          type: string
        count:
          type: number
        percent:
          type: number
        hotelId:
          type: ["number", "null"]
...