# lib/api-zod/src/generated/types/complaintCategory.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hotel Review Insights API
 * OpenAPI spec version: 0.1.0
 */

export interface ComplaintCategory {
  category: string;
  count: number;
  percent: number;
  /** @nullable */
  hotelId?: number | null;
}