# lib/api-zod/src/generated/types/sentimentCount.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Hotel Review Insights API
 * OpenAPI spec version: 0.1.0
 */

export interface SentimentCount {
  hotelId: number;
  hotelName: string;
  positive: number;
  neutral: number;
  negative: number;
}